requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
distro==1.9.0
eval_type_backport==0.2.2
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
mistralai==1.8.2
openai==1.92.3
packaging==25.0
playwright==1.53.0
playwright-stealth==2.0.0
pluggy==1.6.0
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
